requirements.txt
Django==1.6.5
mongoengine==0.8.7
Pillow==2.5.1
oauthlib==0.6.1
pymongo==2.6.3
python-social-auth==0.1.23
python3-openid==3.0.4
requests==2.2.1
requests-oauthlib==0.3.1
six==1.6.1
virtualenv==1.11.6